src/utils/common.py
import os
from box.exceptions import BoxValueError
import yaml
from loguru import logger
import json
import joblib
from box import Box
from pathlib import Path
from typing import Any

def read_yaml(path_to_yaml: Path) -> Box:
    """reads yaml file and returns

    Args:
        path_to_yaml (str): path like input

    Raises:
        ValueError: if yaml file is empty
        e: empty file

    Returns:
        ConfigBox: ConfigBox type
    """
    try:
        with open(path_to_yaml) as yaml_file:
            content = yaml.safe_load(yaml_file)
            logger.info(f"yaml file: {path_to_yaml} loaded successfully")
            return Box(content)
    except BoxValueError:
        raise ValueError("yaml file is empty")
    except Exception as e:
        raise e