requirements.txt
pathlib==1.0.1
numpy==1.22.4
pandas==2.1.4
loguru==0.7.2
python-box==7.3.0
psycopg2-binary==2.9.10
fastparquet==2024.11.0
pyarrow==18.1.0
mlflow==2.19.0
optuna==4.1.0